+++ src/lib/domain/student/index.ts
@@ -1,9 +1,7 @@
 import * as student_interfaces from './interfaces';
 import * as entities from './entities'
 import StudentService from './services/student_service';
-import StudentRepository from './repositories/student_repository';
-import UniversityRepository from './repositories/university_repository';
-import StudyProgramRepository from './repositories/studyprogram_repository';
+import StudentRepository from '../../infrastructure/repositories/student_repository';
 import {S_TYPES} from './types'
 
 
@@ -13,7 +11,5 @@ export {
     student_interfaces,
     entities,
     StudentService,
-    StudentRepository,
-    StudyProgramRepository,
-    UniversityRepository,
+    StudentRepository
 }
