+++ src/lib/infrastructure/index.ts
@@ -1,0 +1,6 @@
+import * as repositories from './repositories'
+
+
+export {
+    repositories
+}
