+++ src/lib/infrastructure/repositories/index.ts
@@ -1,0 +1,9 @@
+import StudentRepository from './student_repository'
+import StudyProgramRepository from './studyprogram_repository'
+import UniversityRepository from './university_repository'
+
+export {
+    StudentRepository,
+    StudyProgramRepository,
+    UniversityRepository
+}
